+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-08-09
+## [Unreleased] - 2022-08-11
 
 ### Added
 * Crossover operators for permutations:
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Uniform Order-Based Crossover (UOBX)
   * Partially Matched Crossover (PMX)
   * Uniform Partially Matched Crossover (UPMX)
+  * Precedence Preservative Crossover (PPX): two-point version
 
 ### Changed